[story/VSPC-232] added factory to create about page form
