+++ vspace/src/main/java/edu/asu/diging/vspace/core/services/IExhibitionAboutPageManager.java
@@ -27,6 +27,4 @@ public interface IExhibitionAboutPageManager {
 
     ExhibitionAboutPage getExhibitionAboutPage();
 
-    AboutPageForm createAboutPageForm();
-
 }
