+++ vspace/src/main/java/edu/asu/diging/vspace/core/services/impl/ExhibitionAboutPageManager.java
@@ -137,85 +137,6 @@ public class ExhibitionAboutPageManager implements IExhibitionAboutPageManager{
                 }
             }        
         }
-    }
-       
-    /**
-     * Creates About Page form object
-     */
-    @Override
-    public AboutPageForm createAboutPageForm() {
-        ExhibitionAboutPage exhibitionAboutPage = getExhibitionAboutPage(); 
-
-        AboutPageForm aboutPageForm=new AboutPageForm();
-        aboutPageForm.setAboutPageText(exhibitionAboutPage.getAboutPageText());
-        aboutPageForm.setTitle(exhibitionAboutPage.getTitle());        
-        
-        IExhibition startExhibtion = exhibitionManager.getStartExhibition();
-        
-        
-        
-        IExhibitionLanguage defaultLanguage = exhibitionManager.getDefaultLanguage(startExhibtion);
-                
-               
-        
-        aboutPageForm.getTitles().add(createLocalizedTitleForm(exhibitionAboutPage, defaultLanguage));
-        
-        aboutPageForm.getAboutPageTexts().add(createLocalizedAboutTextForm(exhibitionAboutPage, defaultLanguage));
-        
-        startExhibtion.getLanguages().forEach(language -> {
-            if(!language.isDefault()) {
-                aboutPageForm.getTitles().add(createLocalizedTitleForm(exhibitionAboutPage, language));
-                
-                aboutPageForm.getAboutPageTexts().add(createLocalizedAboutTextForm(exhibitionAboutPage, language)); 
-            }
-
-
-        });
-        return aboutPageForm;
-    }
-    
-    /**
-     * Creates Localized title object for form 
-     * 
-     * @param exhibitionAboutPage
-     * @param language
-     * @return
-     */
-    private LocalizedTextForm createLocalizedAboutTextForm(ExhibitionAboutPage exhibitionAboutPage,
-            IExhibitionLanguage language) {
-
-        LocalizedTextForm localizedAboutTextForm = new LocalizedTextForm(null, null,  language.getId(), language.getLabel() );
-        ILocalizedText aboutPageText = exhibitionAboutPage.getExhibitionTextDescriptions().stream()
-                .filter(exhibitionText -> StringUtils.equals(language.getId(), exhibitionText.getExhibitionLanguage().getId())).findAny().orElse(null);
-
-        if(aboutPageText != null) {
-            localizedAboutTextForm.setText(aboutPageText.getText());
-            localizedAboutTextForm.setLocalisedTextId( aboutPageText.getId());
-
-        } 
-        localizedAboutTextForm.setIsDefaultExhibitionLanguage(language.isDefault());
-        return localizedAboutTextForm;
-    }
-    
-    /**
-     * 
-     * Creates Localized about text object for form 
-     * @param exhibitionAboutPage
-     * @param language
-     * @return
-     */
-    private LocalizedTextForm createLocalizedTitleForm(ExhibitionAboutPage exhibitionAboutPage, IExhibitionLanguage language) {
-        LocalizedTextForm localizedTitleForm = new LocalizedTextForm(null, null,  language.getId(), language.getLabel() );
-
-        ILocalizedText title = exhibitionAboutPage.getExhibitionTitles().stream()
-                .filter(exhibitionTitle ->  StringUtils.equals(exhibitionTitle.getExhibitionLanguage().getId(), language.getId())).findAny().orElse(null);
-
-        if(title != null) {
-            localizedTitleForm.setText(title.getText());
-            localizedTitleForm.setLocalisedTextId(title.getId());
-        } 
-        localizedTitleForm.setIsDefaultExhibitionLanguage(language.isDefault());      
-        
-        return localizedTitleForm;
-    }
+    } 
+   
 }
