+++ vspace/src/main/java/edu/asu/diging/vspace/web/staff/ExhibitionAboutPageController.java
@@ -16,6 +16,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import org.springframework.web.servlet.view.RedirectView;
 
+import edu.asu.diging.vspace.core.factory.IExhibitionAboutPageFactory;
 import edu.asu.diging.vspace.core.model.IExhibition;
 import edu.asu.diging.vspace.core.model.ILocalizedText;
 import edu.asu.diging.vspace.core.model.impl.ExhibitionAboutPage;
@@ -37,10 +38,13 @@ public class ExhibitionAboutPageController {
 
     @Autowired
     private IExhibitionAboutPageManager aboutPageManager;
+    
+    @Autowired
+    private IExhibitionAboutPageFactory exhibitionAboutPageFactory;
    
     @RequestMapping(value = "/staff/exhibit/about", method = RequestMethod.GET)
     public String showAboutPage(Model model) {    
-        model.addAttribute("aboutPage", aboutPageManager.createAboutPageForm());
+        model.addAttribute("aboutPage", exhibitionAboutPageFactory.createAboutPageForm(aboutPageManager.getExhibitionAboutPage()));
         return "staff/exhibit/aboutPage";
     }
